rag_system/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.0.332
langchain-community==0.0.28
langgraph==0.0.43
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
faiss-cpu==1.7.4
pypdf==3.17.1
unstructured==0.12.2
rank-bm25==0.2.2
nltk==3.8.1
pyppeteer==1.0.2
autoawq==0.1.7
python-multipart==0.0.6
python-dotenv==1.0.0